% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaDaAn-package.R
\docType{package}
\name{BaDaAn-package}
\alias{BaDaAn}
\alias{BaDaAn-package}
\title{BaDaAn: Basic data analysis for minicipality dataset from kolada}
\description{
This package is used to create a linear model by supplying the formula and dataset. The model is calculated using least squares and returns an object of class RC. Within the object you can call the following methods to analyze your model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wountoto/Lab5_AdvR}
}

}
\author{
\strong{Maintainer}: Johannes Hedström \email{johed883@student.liu.se}

Authors:
\itemize{
  \item Mikael Montén \email{mikmo937@student.liu.se}
}

}
\keyword{internal}
