% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dl.R
\name{data_dl}
\alias{data_dl}
\title{Data downloader}
\source{
\url{https://www.kolada.se/?focus=16585}
}
\usage{
data_dl()
}
\value{
The function returns a data frame with 4 columns and 440 rows
}
\description{
Downloading data with an api to kolada, variables and years are 
chosen and translated to English
}
\examples{
# example of how to use the function
df <- data_dl()

}
