% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muniCompare.R
\name{muniCompare}
\alias{muniCompare}
\title{Compare municipalities with a plot}
\usage{
muniCompare(muni_vec, df, var)
}
\arguments{
\item{muni_vec}{A character vector of municipalities in the data frame}

\item{df}{The data frame downloaded with data_dl()}

\item{var}{A string of a KPI variable in the data frame}
}
\value{
A time serie plot of the chosen municipalities(muni_vec) for a variable
from the data frame
}
\description{
Compare municipalities with a plot
}
\examples{
# This is en example of how to use the function
df <- data_dl() # download the data first
muniCompare(c('uppsala','stockholm'), df,'residents')

}
