% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muniCorr.R
\name{muniCorr}
\alias{muniCorr}
\title{Correlation analysis of the municipal spends}
\usage{
muniCorr(df)
}
\arguments{
\item{df}{The dataset which includes variables to be analyzed}
}
\value{
A table and a plot showing correlations between variables.
}
\description{
Performs a correlation analysis for the total municipal spend over the time period by providing a correlation table and scatterplot.
All variables are included in the printout.
}
