% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muniLines.R
\name{muniLines}
\alias{muniLines}
\title{Line plot showing magnitude of all municipal spend}
\usage{
muniLines(muni, df)
}
\arguments{
\item{muni}{A character vector denoting what municipality is to be analyzed}

\item{df}{A dataset containing the municipality of interest}
}
\value{
A line plot showing the spending for the time period
}
\description{
Visualizes municipal spend over the last 10 years for a chosen municipality via lineplot.
Different type of spends are colored accordingly.
}
