% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muniModel.R
\name{muniModel}
\alias{muniModel}
\title{Linear model creation and analyse}
\usage{
muniModel(y, form, df)
}
\arguments{
\item{y}{The independent variable to model for}

\item{form}{The dependent variables to explain the independent variable}

\item{df}{A dataset containing the supplied variables}
}
\value{
A summary printout and distribution plot
}
\description{
Used to create a linear model and subsequently return a summary printout and distribution plot for the model.
The object can handle both simple and multiple linear regressions and handle both numeric and character variables.'
}
