% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaDaAn-package.R
\docType{package}
\name{BaDaAn-package}
\alias{BaDaAn}
\alias{BaDaAn-package}
\title{BaDaAn: Basic Data Analysis for municipality data from Kolada}
\description{
This package serves as an introduction to basic data analysis. Data is based on an API from Kolada involving data over municipal spend for year 2013-2022 in Swedens four largest municipalities - Stockholm, Göteborg, Malmö, and Uppsala. The analysis starts with visualizing how the expenditure magnitudes have changed over time by providing the user with two forms of line plots. One line plot shows all spend variables for one municipality, and one line plot compares municipalities for a chosen variable. The package also provides a function to analyze correlations between variables both by a table and a scatter plot. The final functionality is the ability to create a linear model from chosen variables and return a summary printout as well as distribution plot of the model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wountoto/Lab5_AdvR}
}

}
\author{
\strong{Maintainer}: Johannes Hedström \email{johed883@student.liu.se}

Authors:
\itemize{
  \item Mikael Montén \email{mikmo937@student.liu.se}
}

}
\keyword{internal}
